tests/wpt/tests/css/css-overflow/parsing/line-clamp-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overflow: parsing line-clamp with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow/#propdef-line-clamp">
<meta name="assert" content="line-clamp supports only the grammar 'none | <integer> <‘block-ellipsis’>?'.">
<meta name="assert" content="Zero or negative max-lines integers are invalid.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("line-clamp", 'auto');

test_invalid_value("line-clamp", '0');
test_invalid_value("line-clamp", '-5');

test_invalid_value("line-clamp", '" etc., etc. "');

test_invalid_value("line-clamp", 'none 2');
test_invalid_value("line-clamp", '3 none');
test_invalid_value("line-clamp", 'auto 4');
</script>
</body>
</html>
